chore: move ECS resources into separate file
